Allow setting the jwks-file conditionally within the service-template based on the specified environment.
This requires changing the path of  the default jwks-file as well as adding an additional item within the config map. Instead of setting the flag explicitly, which would lead to overriding environment specific values, we let the
environment be responsible for setting the jwks-file. This way we ensure in production we won't include the static token JWKS.