Add script to pull secrets and terraform data plane clusters (#304)
This script resolves the problem of ensuring we are pulling the correct secrets for the correct environment / data plane cluster. It is idempotent and has no feedback loop, so it should create a more consistent and stable data plane cluster. It will also enable us to continually update the data plane clusters via CI.

Co-authored-by: Stephan Heßelmann <shesselm@redhat.com>