Directly manage Alertmanager configuration
This will enable us to set up a Slack receiver for alerts that we don't
want going to PagerDuty.

This configuration was pulled directly from the generated version of
this secret in one of our data plane clusters and modified to be
generated by this Helm chart instead.  There is no functional difference
between this config and the one generated by the Observability operator.

The PagerDuty and Dead Man's Snitch secrets are deleted since those
secrets are now used directly in the Alertmanager configuration.