ROX-17083: Make PagerDuty the default receiver
This is part of a larger effort to stop treating stage alerts as
critical alerts in PagerDuty.  The specific changes being made in this
PR:

* Pass the `severity` parameter in the PagerDuty receiver configuration.
  This will be used in production to capture non-critical production
  events in PagerDuty without triggering a critical notification to
  on-call engineers.  (These events are currently dropped in
  Alertmanager)
* Switch to the PagerDuty v2 Events API.  This is required for PD to
  recognize the `severity` parameter.  This is done by switching from
  `service_key` to `routing_key`.
* Updated the key in `terraform_cluster.sh` (and thus AWS Secrets
  Manager) to reflect the routing key change.
* Made PagerDuty the default receiver in the Alertmanager config.  This
  will send non-critical alerts to PagerDuty, treating it as a hub for
  all events across all data plane clusters.

It's implied that the PagerDuty routing key for the stage environment
will be different from the one in production.  This is so the stage
service can be configured to force all incidents to be considered low
priority to avoid paging on-call engineers.